CLI: Implement the "omni help" command
With this change, "omni help commands" will list all available
commands, including the ones not displayed in the short list from
"omni --help". Using "omni help <command>" will show the help for
that specific command, which is equivalent to use "omni <command>
--help".

For the moment "omni help topics" returns an empty list of topics,
as there are none yet, and there is no implementation to read the
text files which will cover the help topics. This is to be added
later.